+++ R/test-helpers.R
@@ -5,8 +5,13 @@
 #' @param reserved_func The function under test. This should return `TRUE` when
 #'   an [`ip_address`] is within the reserved network.
 #' @param network The reserved network
+#' @param ignore_before Do not check if address before network is unreserved.
+#'   Used when reserved blocks are contiguous.
+#' @param ignore_after Do not check if address after network is unreserved.
+#'   Used when reserved blocks are contiguous.
 #' @noRd
-expect_reserved_address_range <- function(reserved_func, network) {
+expect_reserved_address_range <- function(reserved_func, network,
+                                          ignore_before = FALSE, ignore_after = FALSE) {
   act <- quasi_label(enquo(reserved_func), arg = "reserved_func")
 
   block_start <- network_address(network)
@@ -15,17 +20,17 @@ expect_reserved_address_range <- function(reserved_func, network) {
   space_end <- ip_address(c("255.255.255.255", "ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff"))
 
   boundaries <- vec_c(
-    if (all(block_start != space_start)) block_start - 1,
+    if (!(ignore_before || any(block_start == space_start))) block_start - 1,
     block_start,
     block_end,
-    if (all(block_end != space_end)) block_end + 1
+    if (!(ignore_after || any(block_end == space_end))) block_end + 1
   )
 
   expected <- c(
-    if (all(block_start != space_start)) FALSE,
+    if (!(ignore_before || any(block_start == space_start))) FALSE,
     TRUE,
     TRUE,
-    if (all(block_end != space_end)) FALSE
+    if (!(ignore_after || any(block_end == space_end))) FALSE
   )
 
   expect(
@@ -43,19 +48,22 @@ expect_reserved_address_range <- function(reserved_func, network) {
 #' @param reserved_func The function under test. This should return `TRUE` when
 #'   an [`ip_network`] is within the reserved network.
 #' @param network The reserved network
+#' @param ignore_super Do not check if supernet is unreserved. Used when
+#'   broadcast address of supernet is reserved.
 #' @noRd
-expect_reserved_network <- function(reserved_func, network) {
+expect_reserved_network <- function(reserved_func, network,
+                                    ignore_super = FALSE) {
   act <- quasi_label(enquo(reserved_func), arg = "reserved_func")
 
   other_networks <- vec_c(
     network,
     if (prefix_length(network) != max_prefix_length(network)) subnets(network),
-    if (prefix_length(network) != 0) supernet(network)
+    if (!(ignore_super || prefix_length(network) == 0)) supernet(network)
   )
   expected <- c(
     TRUE,
     if (prefix_length(network) != max_prefix_length(network)) c(TRUE, TRUE),
-    if (prefix_length(network) != 0) FALSE
+    if (!(ignore_super || prefix_length(network) == 0)) FALSE
   )
 
   expect(
@@ -70,10 +78,12 @@ expect_reserved_network <- function(reserved_func, network) {
 #'
 #' @param reserved_func The function under test. This should return `TRUE` when
 #'   an [`ip_address`] or [`ip_network`] is within the reserved network.
-#' @param network The reserved network
+#' @inheritParams expect_reserved_address_range
+#' @inheritParams expect_reserved_network
 #' @noRd
-expect_reserved <- function(reserved_func, network) {
+expect_reserved <- function(reserved_func, network, ignore_before = FALSE,
+                            ignore_after = FALSE, ignore_super = FALSE) {
   func <- enquo(reserved_func)
-  expect_reserved_address_range(!!func, network)
-  expect_reserved_network(!!func, network)
+  expect_reserved_address_range(!!func, network, ignore_before, ignore_after)
+  expect_reserved_network(!!func, network, ignore_super)
 }