+++ src/IpNetworkVector.cpp
@@ -4,7 +4,7 @@
 #include "masking.h"
 #include "iterative.h"
 #include "sample.h"
-#include "utils.h"
+#include "reserved.h"
 
 using namespace Rcpp;
 
@@ -520,6 +520,13 @@ LogicalVector IpNetworkVector::isLinkLocal() const {
   return checkCondition(
     [](const asio::ip::address_v4 &x) { return (x.to_uint() & 0xFFFF0000) == 0xA9FE0000; },
     [](const asio::ip::address_v6 &x) { return x.is_link_local(); }
+  );
+}
+
+LogicalVector IpNetworkVector::isSiteLocal() const {
+  return checkCondition(
+    [](const asio::ip::address_v4 &x) { return false; },
+    [](const asio::ip::address_v6 &x) { return x.is_site_local(); }
   );
 }